Various fixes
* Updated marker_type value to be the readable name
* Updated the account balance percentage to be a valid number rather than scientific notation
* Removed test api for PBC block api

+++ service/src/main/kotlin/io/provenance/explorer/domain/models/explorer/AssetModels.kt
@@ -28,7 +28,7 @@ data class AssetDetail(
 data class AssetHolder(
     val ownerAddress: String,
     val balance: BigDecimal,
-    val percentage: Double
+    val percentage: BigDecimal
 )
 
 data class AccountDetail(
